Prevent crashes due to "negative" numbers being added to the sound buffer pointer.
It seems that in some cases cc can be less than the counter.
On 64 bit Windows, this is very bad as unsigned long is 32 bit, and pointers are 64 bit.
So a negative number there is really just a really large positive number being added.
This ends up just leading to bad memory accesses when writing to the sound buffer.